fix: 移植上游 EOF 死循环/大PTS/FFmpeg7/RTLD 等修复
从 dmlc/decord open issues/PR 筛选并本地验证后移植：
- NDArray.pts int->int64_t（#269/#341）；SkipFramesImpl 用 Drained() 判真 EOF，避免容器虚标帧数时死循环并采样 EOF 幽灵帧（#373/#342）
- CheckKeyFrame EOF 出口 + GetFrameCount avg_frame_rate.den==0 防除零（#230）
- FFmpeg 7 buffersink pix_fmts 类型变化守卫（libavfilter>=10，PR #368），新增 Windows FFmpeg 7.1 CI job
- 帧索引缓存版本混入 LIBAVFORMAT_VERSION_INT，防止跨 FFmpeg 构建误用旧缓存（本地实测 7.1/8.1 时间戳不一致）
- ctypes 默认 RTLD_LOCAL + DECORD_RTLD_GLOBAL 逃生开关，PATH 缺失兜底（#361/#362/#357）
- _init_internal_api 显式 import _api_internal（PR #364）；AVReader sample_rate 默认值与文档对齐为 -1（PR #237）
- VideoReader.close()/上下文管理器（#222）；GPU VP8 映射（#217）；worker 停止后的 Push 可读报错（#321）
- 新增回归：large_pts.mp4 大 PTS get_batch/seek、close/context manager、PATH 缺失、RTLD_GLOBAL 开关、AVReader 默认值
- 本地验证：Windows GPU FFmpeg 8.1 61 tests、CPU FFmpeg 8.1 60 tests、CPU FFmpeg 7.1 60 tests 全绿

+++ include/decord/runtime/ndarray.h
@@ -36,8 +36,9 @@ namespace runtime {
  */
 class NDArray {
  public:
-  // pts of the frame
-  int pts=-1;
+  // pts of the frame (int64: container PTS is 64-bit; int truncates
+  // large-PTS streams and breaks frame reordering — dmlc/decord#269)
+  int64_t pts=-1;
   // internal container type
   struct Container;
   /*! \brief default constructor */
